Rename value_type_inference -> expression_inference; Make stateless
